major: set testID only on the native Modal component (#452)
Potential BREAKING CHANGE for test suites. 

When testing a component that uses `react-native-modal` using the `testID` prop, the `testID` was set in the Modal component and also in the container view leading to a duplicated `testID`.
This made things tricky when using any of the recommended [test libraries](https://reactnative.dev/docs/testing-overview#component-tests) since querying by that `testID` finds two native elements instead of only one.

With this new update, the `testID` is assigned only to the Modal component (since someone using this package will have control over the content of the Modal and can set a different `testID` if they like). 

Thanks @ssaornil for the fix!